Summary:
Ref T633
see: http://phabricator.n.miliao.com/D1286
 Fix a small bug: version should be byte, for INSECURE_VERSIONS is byte set and
 INSECURE_VERSIONS.contains(version) will return wrong

Test Plan: N/A

Reviewers: xieliang, cuijianwei, fenghonghua

Reviewed By: xieliang

CC: infra-dev

Maniphest Tasks: T633

Differential Revision: http://phabricator.n.miliao.com/D2310

+++ security/src/main/java/org/apache/hadoop/hbase/ipc/SecureServer.java
@@ -407,7 +407,7 @@ public abstract class SecureServer extends HBaseServer {
           if (count < 0 || rpcHeaderBuffer.remaining() > 0) {
             return count;
           }
-          int version = rpcHeaderBuffer.get(0);
+          byte version = rpcHeaderBuffer.get(0);
           byte[] method = new byte[] {rpcHeaderBuffer.get(1)};
           authMethod = AuthMethod.read(new DataInputStream(
               new ByteArrayInputStream(method)));
